algorithms/flexora/requirements.txt
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
datasets>=2.14.0
pyyaml
numpy
tqdm
bitsandbytes>=0.39.0